third_party/WebKit/LayoutTests/fast/encoding/api/replacement-encoding.html

<!DOCTYPE html>
<title>Encoding API: replacement encoding</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="resources/shared.js"></script>
<script>

test(function() {
    assert_throws({name: 'TypeError'}, function() { new TextEncoder('replacement'); });
    assert_throws({name: 'TypeError'}, function() { new TextDecoder('replacement'); });
}, 'The "replacement" label should not be a known encoding.');

encodings_table.forEach(function(section) {
    section.encodings.filter(function(encoding) {
        return encoding.name === 'replacement';
    }).forEach(function(encoding) {
        encoding.labels.forEach(function(label) {
            test(function() {
                assert_throws({name: 'TypeError'}, function() { new TextEncoder(label); });
                assert_throws({name: 'TypeError'}, function() { new TextDecoder(label); });
            }, 'Label for "replacement" should be rejected by API: ' + label);
        });
    });
});

</script>
